Develop To Merge

See merge request s10-webmobile2-sub2/S10P12A106!85

+++ BackEnd/gongtong/src/main/java/site/gongtong/cafe/repository/CafeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import site.gongtong.cafe.model.Cafe;
 
 @Repository
-public interface CafeRepository extends JpaRepository<Cafe,Long> {
+public interface CafeRepository extends JpaRepository<Cafe, Integer> {
 }
